+++ tests/ids-week-view/ids-week-view.spec.ts
@@ -1,7 +1,7 @@
 import AxeBuilder from '@axe-core/playwright';
 import percySnapshot from '@percy/playwright';
-import { Locator, expect } from '@playwright/test';
-import { test } from '../base-fixture';
+import { Locator } from '@playwright/test';
+import { test, expect } from '../base-fixture';
 import IdsWeekView from '../../src/components/ids-week-view/ids-week-view';
 
 test.describe('IdsWeekView tests', () => {
@@ -109,7 +109,218 @@ test.describe('IdsWeekView tests', () => {
     });
 
     test('can set/get startDate', async () => {
-      expect(await idsWeekView.evaluate((element: IdsWeekView) => element.startDate)).toContain('Oct 23 2019');
+      expect(await idsWeekView.evaluate((element: IdsWeekView) => element.startDate)).toBeValidDate();
+      await expect(idsWeekView).toHaveAttribute('start-date');
+
+      expect(await idsWeekView.evaluate((element: IdsWeekView) => {
+        element.startDate = '11/11/2011';
+        return element.startDate;
+      })).toBeValidDate();
+      await expect(idsWeekView).toHaveAttribute('start-date');
+
+      expect(await idsWeekView.evaluate((element: IdsWeekView) => {
+        element.startDate = new Date('2019-10-23T04:00:00.000Z');
+        return element.startDate;
+      })).toBeValidDate();
+      await expect(idsWeekView).toHaveAttribute('start-date');
+
+      expect(await idsWeekView.evaluate((element: IdsWeekView) => {
+        element.startDate = 'Invalid data';
+        return element.startDate;
+      })).toBeValidDate();
+      await expect(idsWeekView).toHaveAttribute('start-date');
+    });
+
+    test('can set/get endDate', async () => {
+      expect(await idsWeekView.evaluate((element: IdsWeekView) => element.endDate)).toBeValidDate();
+      await expect(idsWeekView).toHaveAttribute('end-date');
+
+      expect(await idsWeekView.evaluate((element: IdsWeekView) => {
+        element.endDate = '11/11/2011';
+        return element.endDate;
+      })).toBeValidDate();
+      await expect(idsWeekView).toHaveAttribute('end-date');
+
+      expect(await idsWeekView.evaluate((element: IdsWeekView) => {
+        element.endDate = new Date('2019-10-23T04:00:00.000Z');
+        return element.endDate;
+      })).toBeValidDate();
+      await expect(idsWeekView).toHaveAttribute('end-date');
+
+      expect(await idsWeekView.evaluate((element: IdsWeekView) => {
+        element.endDate = 'Invalid data';
+        return element.endDate;
+      })).toBeValidDate();
+      await expect(idsWeekView).toHaveAttribute('end-date');
+    });
+
+    test('can set/get startHour', async () => {
+      const defHour = 7;
+      const testData = [
+        { data: 9, expected: 9 },
+        { data: '25', expected: defHour },
+        { data: '10', expected: 10 },
+        { data: null, expected: defHour },
+        { data: 8.5, expected: 8.5 },
+        { data: '', expected: defHour },
+        { data: 0, expected: 0 },
+        { data: '-1', expected: defHour }
+      ];
+
+      expect(await idsWeekView.evaluate((element: IdsWeekView) => element.startHour)).toEqual(defHour);
+      await expect(idsWeekView).not.toHaveAttribute('start-hour');
+
+      for (const data of testData) {
+        expect(await idsWeekView.evaluate((element: IdsWeekView, tData) => {
+          element.startHour = tData as any;
+          return element.startHour;
+        }, data.data)).toEqual(data.expected);
+        if (data.data !== null
+          && !Number.isNaN(parseFloat(data.data.toString()))
+          && parseFloat(data.data.toString()) <= 24
+          && parseFloat(data.data.toString()) >= 0
+        ) {
+          await expect(idsWeekView).toHaveAttribute('start-hour', data.expected.toString());
+        } else {
+          await expect(idsWeekView).not.toHaveAttribute('start-hour');
+        }
+      }
+    });
+
+    test('can set/get endHour', async () => {
+      const defHour = 19;
+      const testData = [
+        { data: 9, expected: 9 },
+        { data: '25', expected: defHour },
+        { data: '10', expected: 10 },
+        { data: null, expected: defHour },
+        { data: 8.5, expected: 8.5 },
+        { data: '', expected: defHour },
+        { data: 0, expected: 0 },
+        { data: '-1', expected: defHour }
+      ];
+
+      expect(await idsWeekView.evaluate((element: IdsWeekView) => element.endHour)).toEqual(defHour);
+      await expect(idsWeekView).not.toHaveAttribute('end-hour');
+
+      for (const data of testData) {
+        expect(await idsWeekView.evaluate((element: IdsWeekView, tData) => {
+          element.endHour = tData as any;
+          return element.endHour;
+        }, data.data)).toEqual(data.expected);
+        if (data.data !== null
+          && !Number.isNaN(parseFloat(data.data.toString()))
+          && parseFloat(data.data.toString()) <= 24
+          && parseFloat(data.data.toString()) >= 0
+        ) {
+          await expect(idsWeekView).toHaveAttribute('end-hour', data.expected.toString());
+        } else {
+          await expect(idsWeekView).not.toHaveAttribute('end-hour');
+        }
+      }
+    });
+
+    test('can set/get showTimeline', async () => {
+      const testData = [
+        { data: true, expected: true },
+        { data: 'false', expected: false },
+        { data: '', expected: true },
+        { data: null, expected: false }
+      ];
+
+      expect(await idsWeekView.evaluate((element: IdsWeekView) => element.showTimeline)).toBeFalsy();
+      await expect(idsWeekView).toHaveAttribute('show-timeline', 'false');
+
+      for (const data of testData) {
+        expect(await idsWeekView.evaluate((element: IdsWeekView, tData) => {
+          element.showTimeline = tData as any;
+          return element.showTimeline;
+        }, data.data)).toEqual(data.expected);
+        if (data.expected) {
+          await expect(idsWeekView).toHaveAttribute('show-timeline', 'true');
+        } else {
+          await expect(idsWeekView).toHaveAttribute('show-timeline', 'false');
+        }
+      }
+    });
+
+    test('can set/get timelineInterval', async () => {
+      const defTime = 30000;
+      const testData = [
+        { data: '1000', expected: 1000 },
+        { data: 0, expected: defTime },
+        { data: 3000, expected: 3000 },
+        { data: null, expected: defTime },
+        { data: '5.5', expected: 5.5 },
+        { data: '', expected: defTime }
+      ];
+
+      expect(await idsWeekView.evaluate((element: IdsWeekView) => element.timelineInterval)).toEqual(defTime);
+      await expect(idsWeekView).not.toHaveAttribute('timeline-interval');
+
+      for (const data of testData) {
+        expect(await idsWeekView.evaluate((element: IdsWeekView, tData) => {
+          element.timelineInterval = tData as any;
+          return element.timelineInterval;
+        }, data.data)).toEqual(data.expected);
+        if (data.data !== null
+          && !Number.isNaN(parseFloat(data.data.toString()))
+          && parseFloat(data.data.toString()) > 0
+        ) {
+          await expect(idsWeekView).toHaveAttribute('timeline-interval', data.expected.toString());
+        } else {
+          await expect(idsWeekView).not.toHaveAttribute('timeline-interval');
+        }
+      }
+    });
+
+    test('can trigger calendar event\'s events', async ({ eventsTest }) => {
+      const eventItems = [
+        {
+          id: '1',
+          subject: 'Intraday Event',
+          starts: '2019-10-23T12:00:00.000',
+          ends: '2019-10-23T12:15:00.000',
+          type: 'dto',
+          isAllDay: 'false'
+        },
+        {
+          id: '2',
+          subject: 'All Day Event',
+          starts: '2019-10-24T00:00:00.000',
+          ends: '2019-10-24T23:59:59.999',
+          type: 'admin',
+          isAllDay: 'true'
+        }
+      ];
+
+      const eventTypes = [
+        {
+          id: 'dto',
+          label: 'Discretionary Time Off',
+          translationKey: 'DiscretionaryTimeOff',
+          color: 'blue',
+          checked: true
+        },
+        {
+          id: 'admin',
+          label: 'Admin',
+          translationKey: 'AdministrativeLeave',
+          color: 'purple',
+          checked: true
+        }
+      ];
+      const handle = (await idsWeekView.elementHandle())!;
+      await eventsTest.onEvent('ids-week-view', 'beforeeventrendered', handle);
+      await eventsTest.onEvent('ids-week-view', 'aftereventrendered', handle);
+
+      await idsWeekView.evaluate((element: IdsWeekView, data: any) => {
+        element.eventTypesData = data.eventTypes;
+        element.eventsData = data.eventItems;
+      }, { eventItems, eventTypes });
+
+      expect(await eventsTest.isEventTriggered('ids-week-view', 'beforeeventrendered')).toBeTruthy();
+      expect(await eventsTest.isEventTriggered('ids-week-view', 'aftereventrendered')).toBeTruthy();
     });
   });
 });